Add tracing (OpenTelemetry)
This commit instruments thanos-remote-read with OpenTelemetry tracing. It uses
the library provided by OpenTelemetry and as of now, uses the Jaeger exporter
for compatibility.

This implementation adds the tracing integration to gRPC and the HTTP server
aswell as adding the most basic span to each function. The aim is to add
multiple child spam to provide greater detail in the future.